+++ xray/core/dataset.py
@@ -17,6 +17,7 @@ from . import variable
 from . import alignment
 from . import formatting
 from .. import backends, conventions
+from .alignment import align, partial_align
 from .coordinates import DatasetCoordinates, Indexes
 from .common import ImplementsDatasetReduce, AttrAccessMixin
 from .utils import Frozen, SortedKeysDict, ChainMap
@@ -171,7 +172,18 @@ def _as_dataset_variable(name, var):
     return var
 
 
-def _expand_arrays(raw_variables, old_variables={}, compat='identical'):
+def _align_variables(arrays, join='outer'):
+    """Align all DataArrays in the provided dict, leaving other values alone.
+    """
+    alignable = [k for k, v in arrays.items() if hasattr(v, 'indexes')]
+    aligned = align(*[arrays[a] for a in alignable],
+                    join=join, copy=False)
+    new_arrays = OrderedDict(arrays)
+    new_arrays.update(zip(alignable, aligned))
+    return new_arrays
+
+
+def _expand_variables(raw_variables, old_variables={}, compat='identical'):
     """Expand a dictionary of variables.
 
     Returns a dictionary of Variable objects suitable for inserting into a
@@ -189,19 +201,21 @@ def _expand_arrays(raw_variables, old_variables={}, compat='identical'):
     variables = ChainMap(new_variables, old_variables)
 
     def add_variable(name, var):
+        var = _as_dataset_variable(name, var)
         if name not in variables:
-            variables[name] = _as_dataset_variable(name, var)
+            variables[name] = var
+            new_coord_names.update(variables[name].dims)
         else:
             if not getattr(variables[name], compat)(var):
                 raise ValueError('conflicting value for variable %s:\n'
                                  'first value: %r\nsecond value: %r'
                                  % (name, variables[name], var))
             if compat == 'broadcast_equals':
-                new_dims = _as_dataset_variable(name, var).dims
                 common_dims = OrderedDict(zip(variables[name].dims,
                                               variables[name].shape))
                 common_dims.update(zip(var.dims, var.shape))
                 variables[name] = variables[name].set_dims(common_dims)
+                new_coord_names.update(var.dims)
 
     for name, var in iteritems(raw_variables):
         if hasattr(var, 'coords'):
@@ -209,8 +223,10 @@ def _expand_arrays(raw_variables, old_variables={}, compat='identical'):
             new_coord_names.update(var.coords)
             for dim, coord in iteritems(var.coords):
                 if dim != name:
-                    add_variable(dim, coord)
+                    add_variable(dim, coord.variable)
+            var = var.variable
         add_variable(name, var)
+
     return new_variables, new_coord_names
 
 
@@ -238,19 +254,38 @@ def _calculate_dims(variables):
     return dims
 
 
-class _DatasetLike(object):
-    """A Dataset-like object that only contains a few private attributes
+def _merge_expand(aligned_self, other, overwrite_vars, compat):
+    possible_conflicts = dict((k, v) for k, v in aligned_self._arrays.items()
+                              if k not in overwrite_vars)
+    new_vars, new_coord_names = _expand_variables(other, possible_conflicts, compat)
+    replace_vars = aligned_self._arrays.copy()
+    replace_vars.update(new_vars)
+    return replace_vars, new_vars, new_coord_names
 
-    Like `as_dataset`, handles DataArrays, Datasets and dictionaries of
-    variables. The difference is that this method never creates a new Dataset
-    object, and hence is much more lightweight, avoiding any consistency
-    checks on the variables (that should be handled later).
-    """
-    def __init__(self, obj):
-        obj = getattr(obj, '_dataset', obj)
-        self._arrays = getattr(obj, '_arrays', obj)
-        self._coord_names = getattr(obj, '_coord_names', set())
-        self.attrs = getattr(obj, 'attrs', {})
+
+def _merge_dataset(self, other, overwrite_vars, compat, join):
+    aligned_self, other = partial_align(self, other, join=join, copy=False)
+
+    replace_vars, new_vars, new_coord_names = _merge_expand(
+        aligned_self, other._arrays, overwrite_vars, compat)
+    new_coord_names.update(other._coord_names)
+
+    return replace_vars, new_vars, new_coord_names
+
+
+def _merge_dict(self, other, overwrite_vars, compat, join):
+    other = _align_variables(other, join='outer')
+
+    alignable = [k for k, v in other.items() if hasattr(v, 'indexes')]
+    aligned = partial_align(self, *[other[a] for a in alignable],
+                            join=join, copy=False, exclude=overwrite_vars)
+
+    aligned_self = aligned[0]
+
+    other = OrderedDict(other)
+    other.update(zip(alignable, aligned[1:]))
+
+    return _merge_expand(aligned_self, other, overwrite_vars, compat)
 
 
 def _assert_empty(args, msg='%s'):
@@ -323,7 +358,8 @@ class Dataset(Mapping, ImplementsDatasetReduce, AttrAccessMixin):
     _attrs = None
     _arrays = Frozen({})
 
-    def __init__(self, variables=None, coords=None, attrs=None):
+    def __init__(self, variables=None, coords=None, attrs=None,
+                 compat='broadcast_equals'):
         """To load data from a file or file-like object, use the `open_dataset`
         function.
 
@@ -358,12 +394,12 @@ class Dataset(Mapping, ImplementsDatasetReduce, AttrAccessMixin):
         if coords is None:
             coords = set()
         if variables or coords:
-            self._set_init_vars_and_dims(variables, coords)
+            self._set_init_vars_and_dims(variables, coords, compat)
         if attrs is not None:
             self.attrs = attrs
 
-    def _add_missing_coords(self):
-        """Add missing coordinates IN-PLACE to _arrays
+    def _add_missing_coords_inplace(self):
+        """Add missing coordinates to self._arrays
         """
         for dim, size in iteritems(self.dims):
             if dim not in self._arrays:
@@ -398,25 +434,23 @@ class Dataset(Mapping, ImplementsDatasetReduce, AttrAccessMixin):
         # all checks are complete: it's safe to update
         self._arrays = arrays
         self._dims = dims
-        self._add_missing_coords()
-        self._coord_names.update(dims)
+        self._add_missing_coords_inplace()
         self._coord_names.update(new_coord_names)
 
-    def _set_init_vars_and_dims(self, vars, coords):
+    def _set_init_vars_and_dims(self, vars, coords, compat):
         """Set the initial value of Dataset arrays and dimensions
         """
         _assert_empty([k for k in vars if k in coords],
                       'redundant variables and coordinates: %s')
         arrays = ChainMap(vars, coords)
 
-        new_arrays, new_coord_names = _expand_arrays(arrays)
-        _assert_empty([k for k in new_coord_names if k not in new_arrays],
-                      'no matching variables exist for some coordinates: %s')
+        aligned = _align_variables(arrays)
+        new_variables, new_coord_names = _expand_variables(aligned,
+                                                           compat=compat)
 
         new_coord_names.update(coords)
-        self._update_vars_and_coords(new_arrays, new_coord_names,
-                                     needs_copy=False,
-                                     check_coord_names=False)
+        self._update_vars_and_coords(new_variables, new_coord_names,
+                                     needs_copy=False, check_coord_names=False)
 
     @classmethod
     def load_store(cls, store, decoder=None):
@@ -681,7 +715,7 @@ class Dataset(Mapping, ImplementsDatasetReduce, AttrAccessMixin):
         if utils.is_dict_like(key):
             raise NotImplementedError('cannot yet use a dictionary as a key '
                                       'to set Dataset values')
-        self.merge({key: value}, inplace=True, overwrite_vars=[key])
+        self.update({key: value})
 
     def __delitem__(self, key):
         """Remove a variable from this dataset.
@@ -1078,8 +1112,7 @@ class Dataset(Mapping, ImplementsDatasetReduce, AttrAccessMixin):
         return obj
 
     def update(self, other, inplace=True):
-        """Update this dataset's variables and attributes with those from
-        another dataset.
+        """Update this dataset's variables with those from another dataset.
 
         Parameters
         ----------
@@ -1100,14 +1133,11 @@ class Dataset(Mapping, ImplementsDatasetReduce, AttrAccessMixin):
             If any dimensions would have inconsistent sizes in the updated
             dataset.
         """
-        other = _DatasetLike(other)
-        obj = self.merge(other, inplace=inplace,
-                         overwrite_vars=other._arrays)
-        obj.attrs.update(other.attrs)
-        return obj
+        return self.merge(
+            other, inplace=inplace, overwrite_vars=list(other), join='left')
 
     def merge(self, other, inplace=False, overwrite_vars=set(),
-              compat='broadcast_equals'):
+              compat='broadcast_equals', join='outer'):
         """Merge the arrays of two datasets into a single dataset.
 
         This method generally not allow for overriding data, with the exception
@@ -1134,6 +1164,13 @@ class Dataset(Mapping, ImplementsDatasetReduce, AttrAccessMixin):
             - 'equals': all values and dimensions must be the same.
             - 'identical': all values, dimensions and attributes must be the
               same.
+        join : {'outer', 'inner', 'left', 'right'}, optional
+            Method for joining ``self`` and ``other`` along shared dimensions:
+
+            - 'outer': use the union of the indexes
+            - 'inner': use the intersection of the indexes
+            - 'left': use indexes from ``self``
+            - 'right': use indexes from ``other``
 
         Returns
         -------
@@ -1148,34 +1185,26 @@ class Dataset(Mapping, ImplementsDatasetReduce, AttrAccessMixin):
         if compat not in ['broadcast_equals', 'equals', 'identical']:
             raise ValueError("compat=%r invalid: must be 'broadcast_equals', "
                              "'equals' or 'identical'" % compat)
-        other = _DatasetLike(other)
 
-        # determine variables to check for conflicts
-        if not overwrite_vars:
-            potential_conflicts = self._arrays
-        else:
-            if isinstance(overwrite_vars, basestring):
-                overwrite_vars = set([overwrite_vars])
-            else:
-                overwrite_vars = set(overwrite_vars)
-            potential_conflicts = dict((k, v) for k, v in iteritems(self._arrays)
-                                       if k not in overwrite_vars)
+        if isinstance(overwrite_vars, basestring):
+            overwrite_vars = [overwrite_vars]
+        overwrite_vars = set(overwrite_vars)
 
-        new_variables, new_coord_names = _expand_arrays(
-            other._arrays, potential_conflicts, compat)
-        new_coord_names |= other._coord_names
+        merge = _merge_dataset if isinstance(other, Dataset) else _merge_dict
 
-        _assert_empty([k for k in other._arrays
-                       if k in potential_conflicts
-                       and k not in new_coord_names
-                       and k in self.coords],
-                      'variables with these names already exist as '
-                      'coordinates: %s')
+        replace_vars, new_vars, new_coord_names = merge(
+            self, other, overwrite_vars, compat=compat, join=join)
 
-        # update variables
+        newly_coords = new_coord_names & (set(self) - set(self.coords))
+        no_longer_coords = set(self.coords) & (set(new_vars) - new_coord_names)
+        ambiguous_coords = (newly_coords | no_longer_coords) - overwrite_vars
+        if ambiguous_coords:
+            raise ValueError('cannot merge: the following variables are '
+                             'coordinates on one dataset but not the other: %s'
+                             % list(ambiguous_coords))
+
         obj = self if inplace else self.copy()
-        obj._update_vars_and_coords(new_variables, new_coord_names,
-                                    needs_copy=inplace)
+        obj._update_vars_and_coords(replace_vars, new_coord_names)
         return obj
 
     def _assert_all_in_dataset(self, names, virtual_okay=False):
@@ -1651,6 +1680,12 @@ class Dataset(Mapping, ImplementsDatasetReduce, AttrAccessMixin):
         def func(self, other):
             if isinstance(other, groupby.GroupBy):
                 return NotImplemented
+            if hasattr(other, 'indexes'):
+                self, other = align(self, other, join='inner', copy=False)
+                empty_indexes = [d for d, s in self.dims.items() if s == 0]
+                if empty_indexes:
+                    raise ValueError('no overlapping labels for some '
+                                     'dimensions: %s' % empty_indexes)
             other_coords = getattr(other, 'coords', None)
             ds = self.coords.merge(other_coords)
             g = f if not reflexive else lambda x, y: f(y, x)
@@ -1684,15 +1719,23 @@ def _calculate_binary_op(f, dataset, other, dest_vars):
     if utils.is_dict_like(other):
         other_arrays = getattr(other, '_arrays', other)
         other_vars = getattr(other, 'vars', other)
-        if set(dataset_vars) != set(other_vars):
-            raise ValueError('Datasets do not have the same variables: '
-                             '%s, %s' % (list(dataset_vars), list(other_vars)))
+        performed_op = False
         for k in dataset_vars:
-            dest_vars[k] = f(dataset_arrays[k], other_arrays[k])
+            if k in other_vars:
+                dest_vars[k] = f(dataset_arrays[k], other_arrays[k])
+                performed_op = True
+            elif k in dest_vars:
+                # we are doing an in-place operation
+                raise ValueError('datasets must have the same variables for '
+                                 'in-place arithmetic operations: %s, %s'
+                                 % (list(dataset_vars), list(other_vars)))
+        if not performed_op:
+            raise ValueError('datasets have no overlapping variables: %s, %s'
+                             % (list(dataset_vars), list(other_vars)))
     else:
-        other_arrays = getattr(other, 'variable', other)
+        other_variable = getattr(other, 'variable', other)
         for k in dataset_vars:
-            dest_vars[k] = f(dataset_arrays[k], other_arrays)
+            dest_vars[k] = f(dataset_arrays[k], other_variable)
 
 
 ops.inject_all_ops_and_reduce_methods(Dataset, array_only=False)